# Eliminate dynamic calls