Bench results below. Runtime results are probably too noisy and useless.

## AOT runtime

|                      | Before    | After    | Diff
|----------------------|-----------|----------|-
|ToProto3JsonObject    | 10659 μs  | 10533 μs | -126 μs, -1.18%
|HashCode              | 4025 μs   | 3992 μs  | -33 μs, -0.82%
|ToBinary              | 11969 μs  | 12035 μs | +66 μs, +0.55%
|FromProto3JsonString  | 30021 μs  | 29646 μs | -375 μs, -1.25%
|FromJson              | 27979 μs  | 28172 μs | +193 μs, +0.69%
|FromBinary            | 6406 μs   | 6394 μs  | -12 μs, -0.19%
|ToProto3JsonString    | 36102 μs  | 35267 μs | -835 μs, -2.31%
|ToJson                | 33937 μs  | 32986 μs | -951 μs, -2.80%
|FromProto3JsonObject  | 6619 μs   | 6531 μs  | -88 μs, -1.33%

## JS runtime

|                      | Before    | After    | Diff
|----------------------|-----------|----------|-
|ToProto3JsonObject    | 20618 μs  | 19851 μs | -767 μs, -3.72%
|HashCode              | 13698 μs  | 12808 μs | -890 μs, -6.50%
|ToBinary              | 36781 μs  | 36563 μs | -218 μs, -0.59%
|FromProto3JsonString  | 37055 μs  | 34084 μs | -2971 μs, -8.02%
|FromJson              | 47651 μs  | 44755 μs | -2896 μs, -6.08%
|FromBinary            | 41750 μs  | 39392 μs | -2358 μs, -5.65%
|ToProto3JsonString    | 67566 μs  | 71928 μs | +4362 μs, +6.46%
|ToJson                | 73392 μs  | 77807 μs | +4415 μs, +6.02%
|FromProto3JsonObject  | 17084 μs  | 17336 μs | +252 μs, +1.48%

## Benchmark binary size

|    | Before          | After          | Diff
|----|-----------------|----------------|-
|AOT | 7,122,256 bytes | 7,076,424 bytes| -45,832 bytes, -0.64%
|JS  | 347,371 bytes   | 343,904 bytes  | -3,467 bytes, -0.99%

Fixes #578