@osa1 2022-07-13
I benchmarked this PR on Golem: https://golem.corp.goog/Revision?repository=protobuf&revision=84839&patch=17189

The results are mixed. Some results that I found interesting:

- On AOT, `from_binary` regressed by 1.1%, but `query_decode_binary` improved by 0.6%. Both of these are the same benchmark (binary deserialization) on different input, so assuming the results are accurate and not noisy, it seems like this PR makes deserialization of some fields faster and some fields slower.

- On JS, `to_binary` regressed by 3.5%, but `query_encode_binary` improved by 2.8%. This case is the same as above: these are the same benchmarks with different inputs.

- Binary encoding got faster on AOT but slower on JS. Binary decoding got faster on JS but slower on AOT.